feat: implement first draft of creating table entities from url
